# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "12-factor Rust backend: layered store + RBAC + JWT + WS + worker",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/routes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/routes` — list routes (admin), with optional brand\nfilter, search and offset pagination.",
        "operationId": "list",
        "parameters": [
          {
            "name": "brandId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive search over the route name, start/end city\nslugs and the owning brand's name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (clamped to `[1, 200]` by the service). `None` = all\nrows (legacy consumers).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset (0-based).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRouteListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "`POST /api/admin/routes` — create a route.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/routes/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "`PUT /api/admin/routes/{id}` — update a route.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "`DELETE /api/admin/routes/{id}` — delete a route.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/schedules": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/schedules?routeId=` — list schedules for a route.",
        "operationId": "list",
        "parameters": [
          {
            "name": "routeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminScheduleListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "`POST /api/admin/schedules` — create a schedule.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/schedules/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "`PUT /api/admin/schedules/{id}` — update a schedule.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Schedule ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "`DELETE /api/admin/schedules/{id}` — delete a schedule.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Schedule ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/routes": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "`GET /api/routes` — list routes.",
        "operationId": "routes",
        "parameters": [
          {
            "name": "brand_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminAddressOut": {
        "type": "object",
        "required": [
          "id",
          "brandId",
          "name",
          "lat",
          "lon",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "brandId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "district": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "province": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string"
          },
          "ward": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AdminMutationResponse": {
        "type": "object",
        "description": "Response of `POST /api/admin/brands` + `PUT /api/admin/brands/{id}` + `DELETE`.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "AdminPlacePreview": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "City slug (e.g. `\"ha-noi\"`). NOT a UUID — matches the\n`start_location_id` / `end_location_id` column type on `route`."
          },
          "name": {
            "type": "string"
          },
          "province": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AdminRouteListResponse": {
        "type": "object",
        "description": "Response of `GET /api/admin/routes`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminRouteOut"
            }
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Total matching-row count (independent of pagination). Omitted\nwhen the caller didn't paginate (legacy \"fetch all\" consumers).",
            "minimum": 0
          }
        }
      },
      "AdminRouteOut": {
        "type": "object",
        "description": "Admin route list item — like `RouteOut` but with start/end place\npreviews and schedule/pickup-point counts.",
        "required": [
          "id",
          "name",
          "startLocationId",
          "endLocationId",
          "status",
          "createdAt",
          "updatedAt",
          "scheduleCount",
          "pickupPointCount"
        ],
        "properties": {
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "endLocation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminPlacePreview"
              }
            ]
          },
          "endLocationId": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "pickupPointCount": {
            "type": "integer",
            "format": "int64"
          },
          "scheduleCount": {
            "type": "integer",
            "minimum": 0
          },
          "startLocation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminPlacePreview"
              }
            ]
          },
          "startLocationId": {
            "type": "string",
            "description": "City slug (e.g. `\"ha-noi\"`). Required — every route has both a\nstart and an end city. Resolved to `start_location` via\n`crate::cities::find_by_slug`."
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "AdminScheduleListResponse": {
        "type": "object",
        "description": "Response of `GET /api/admin/schedules`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminScheduleOut"
            }
          }
        }
      },
      "AdminScheduleOut": {
        "type": "object",
        "required": [
          "id",
          "routeId",
          "departureTime",
          "basePriceAdult",
          "points",
          "createdAt"
        ],
        "properties": {
          "amenities": {
            "type": [
              "string",
              "null"
            ]
          },
          "basePriceAdult": {
            "type": "integer",
            "format": "int64"
          },
          "basePriceChild": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "busLayoutId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Bus layout id (string on the wire, `Uuid` in Rust — see the\nentity field comment about the old String/TEXT drift)."
          },
          "createdAt": {
            "type": "string"
          },
          "daysOfWeek": {
            "type": [
              "string",
              "null"
            ]
          },
          "departureTime": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": [
              "string",
              "null"
            ]
          },
          "effectiveTo": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSchedulePointOut"
            },
            "description": "Ordered address points (departure → midway stops → destination).\nEmpty when the schedule has no point sequence yet."
          },
          "routeId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminVehicleTypeOut",
                "description": "The resolved vehicle class row (schedule's explicit type; the\nbus-layout fallback is NOT resolved here — the admin UI shows\nthe catalog type it configured)."
              }
            ]
          },
          "vehicleTypeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Explicit vehicle class (`vehicle_type` row). `None` = resolved\nthrough the bus layout (legacy behaviour)."
          }
        }
      },
      "AdminSchedulePointOut": {
        "type": "object",
        "required": [
          "id",
          "scheduleId",
          "addressId",
          "stopOrder",
          "kind",
          "address"
        ],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/AdminAddressOut"
          },
          "addressId": {
            "type": "string",
            "format": "uuid"
          },
          "arrivalTime": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional `HH:MM` — when the vehicle reaches this stop."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "description": "`pickup` (first) / `middle` / `drop` (last)."
          },
          "scheduleId": {
            "type": "string",
            "format": "uuid"
          },
          "stopOrder": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AdminVehicleTypeOut": {
        "type": "object",
        "required": [
          "id",
          "code",
          "label",
          "sortOrder",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable slug (unique, lowercase — matches the legacy\n`bus_layout.vehicle_type` codes and the public search filter)."
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "description": "Display name (Vietnamese)."
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "description": "Display order in pickers (ascending)."
          },
          "status": {
            "type": "string",
            "description": "`active` | `disabled`."
          },
          "totalSeats": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Typical seat count — informational."
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "RouteBrandPreview": {
        "type": "object",
        "description": "Brand preview embedded in `RouteOut`.",
        "properties": {
          "accentColor": {
            "type": [
              "string",
              "null"
            ]
          },
          "logoUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "RouteEndpoint": {
        "type": "object",
        "description": "From/To endpoint embedded in `RouteOut`.",
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "RouteListResponse": {
        "type": "object",
        "description": "Response of `GET /api/routes`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteOut"
            }
          }
        }
      },
      "RouteOut": {
        "type": "object",
        "description": "A route list item returned by `GET /api/routes`.",
        "required": [
          "id",
          "name",
          "brand",
          "from",
          "to",
          "scheduleCount"
        ],
        "properties": {
          "brand": {
            "$ref": "#/components/schemas/RouteBrandPreview"
          },
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "from": {
            "$ref": "#/components/schemas/RouteEndpoint"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "scheduleCount": {
            "type": "integer",
            "minimum": 0
          },
          "to": {
            "$ref": "#/components/schemas/RouteEndpoint"
          }
        }
      },
...
      "UpsertRouteRequest": {
        "type": "object",
        "description": "Request body for `POST /api/admin/routes` + `PUT /api/admin/routes/{id}`.",
        "properties": {
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "endLocationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "startLocationId": {
            "type": [
              "string",
              "null"
            ],
            "description": "City slug (e.g. `\"ha-noi\"`). Max 20 chars — matches the DB\ncolumn `VARCHAR(20) NOT NULL`. The slug MUST be one of the\nvalues in `crate::cities::CITIES`; we don't enforce that here\n(validator doesn't have access to the static list) but the\nfrontend dropdown only sends known slugs. Required — a route\nwithout start/end cities is meaningless."
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UpsertSchedulePointItem": {
        "type": "object",
        "description": "One entry of the ordered `points` array on schedule create/update.\nThe array position defines `stopOrder`; `kind` is derived\n(first = `pickup`, last = `drop`, else `middle`).",
        "properties": {
          "addressId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "arrivalTime": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional `HH:MM` arrival time at this stop. Validated in the\nservice (00:00–23:59); `null`/omitted clears it."
          }
        }
      },
      "UpsertScheduleRequest": {
        "type": "object",
        "description": "Request body for `POST /api/admin/schedules` + `PUT /api/admin/schedules/{id}`.",
        "properties": {
          "amenities": {
            "type": [
              "string",
              "null"
            ]
          },
          "basePriceAdult": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "basePriceChild": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "busLayoutId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Bus layout id (string on the wire, `Uuid` in Rust)."
          },
          "daysOfWeek": {
            "type": [
              "string",
              "null"
            ]
          },
          "departureTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "effectiveFrom": {
            "type": [
              "string",
              "null"
            ]
          },
          "effectiveTo": {
            "type": [
              "string",
              "null"
            ]
          },
          "points": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/UpsertSchedulePointItem"
            },
            "description": "When present, replaces the schedule's whole point sequence.\nRequires ≥ 2 items (departure + destination); `kind` and\n`stopOrder` are derived from the array position."
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "vehicleTypeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Vehicle class from the admin-managed `vehicle_type` catalog.\nValidated in the service (404-style validation error when unknown)."
          }
        }
      },
...
    }
  },
...
}